Rename canister_call.rs into canister.rs test file
